# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "spassu",
        "version": "1"
    },
    "paths": {
        "/api/authors": {
            "get": {
                "tags": [
                    "authors"
                ],
                "summary": "List all Authors",
                "operationId": "9ddd2673eddfc76fec3bc5eab3a55f60",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the author"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Author"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "authors"
                ],
                "summary": "Add new Author",
                "operationId": "30e12a97123554a618e5b40557581705",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddAuthorDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "author": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/authors/{id}": {
            "get": {
                "tags": [
                    "authors"
                ],
                "summary": "Get Author",
                "operationId": "d821d782c8be4226a186d74a818e4bd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the author"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "author": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "authors"
                ],
                "summary": "Update Author",
                "operationId": "375c0cf98fe358196e0739be08513985",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the author"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "author": {
                                            "$ref": "#/components/schemas/Author"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "authors"
                ],
                "summary": "Delete Author",
                "operationId": "6f01db38954df4ab5aaf09b7681477de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the author"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "List books",
                "operationId": "96a05c4c3d4f63a319efc8844a64009a",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "books"
                ],
                "summary": "Add new book",
                "operationId": "9f2fd429a00fccae186fe636793262ce",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddBookDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subject": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "books"
                ],
                "summary": "Remove the book",
                "operationId": "26a97b55d197e37779144274e7a3ccfb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the book"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get book",
                "operationId": "7cabace9f9f77012c39b7394aca585ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the book"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subject": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "books"
                ],
                "summary": "Update book",
                "operationId": "d18d89a3b8a689e2dd62dae1c9301b95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the book"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddBookDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "subject": {
                                            "$ref": "#/components/schemas/Book"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AddAuthorDTO": {
                "title": "AuthUserDTO",
                "properties": {
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AddBookDTO": {
                "title": "AddBookDTO",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "publisher": {
                        "type": "string"
                    },
                    "edition": {
                        "type": "integer"
                    },
                    "publicationYear": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "subjects": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
...
            "Author": {
                "title": "Author",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Book": {
                "title": "Book",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "publisher": {
                        "type": "string"
                    },
                    "edition": {
                        "type": "integer"
                    },
                    "publicationYear": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Author"
                        }
                    },
                    "subjects": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Subject"
                        }
                    }
                },
                "type": "object"
            },
            "Subject": {
                "title": "Subject",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}